+++ src/main/java/manager/controller/FeeController.java
@@ -3,12 +3,10 @@ package manager.controller;
 import lombok.RequiredArgsConstructor;
 import manager.dto.FeeRequestDTO;
 import manager.dto.FeeResponseDTO;
-import manager.entity.Fee;
 import manager.service.FeeService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -22,12 +20,12 @@ public class FeeController {
         return ResponseEntity.ok(feeService.findAll());
     }
     @GetMapping(name = "id")
-    public ResponseEntity<FeeResponseDTO> getFee(@RequestParam BigInteger id) {
+    public ResponseEntity<FeeResponseDTO> getFee(@RequestParam Long id) {
         return ResponseEntity.ok(feeService.findById(id));
     }
 
     @GetMapping(name = "category")
-    public ResponseEntity<List<FeeResponseDTO>> getFeeByRoomCategoryId(@RequestParam BigInteger idCategory) {
+    public ResponseEntity<List<FeeResponseDTO>> getFeeByRoomCategoryId(@RequestParam Integer idCategory) {
         List<FeeResponseDTO> responseDTOS= feeService.findByRoomCategoryId(idCategory);
         return ResponseEntity.ok().body(responseDTOS);
     }
@@ -38,12 +36,12 @@ public class FeeController {
     }
 
     @PutMapping("id/{id}")
-    public ResponseEntity<FeeResponseDTO> updateFee(@RequestBody FeeRequestDTO feeRequestDTO, @PathVariable BigInteger id) {
+    public ResponseEntity<FeeResponseDTO> updateFee(@RequestBody FeeRequestDTO feeRequestDTO, @PathVariable Long id) {
         FeeResponseDTO fee= feeService.update(feeRequestDTO, id);
         return ResponseEntity.ok(fee);
     }
     @DeleteMapping("id/{id}")
-    public ResponseEntity<?> deleteFee(@PathVariable BigInteger id) {
+    public ResponseEntity<?> deleteFee(@PathVariable Long id) {
         feeService.delete(id);
         return ResponseEntity.ok().build();
     }
