+++ src/main/java/manager/dto/ReservationRequestDTO.java
@@ -16,5 +16,5 @@ public class ReservationRequestDTO {
     private Integer reservationHotelId;
     private Integer reservationRoomCategoryId;
 
-    private BigInteger reservationClientId;
+    private Long reservationClientId;
 }
